Function leader::load_dotenvy_vars_if_present

pub(crate) fn load_dotenvy_vars_if_present()

Attempt to load in the local .env if present and set any environment variables specified inside of it.

To keep things simple, any IO error we will treat as the file not existing and continue moving on without the env variables set.